requirements.txt
Django==1.10.1
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.4.7
markdown2==2.3.1
Pillow==3.3.1
